Added support for LMSTYLES16 bspx lump, allowing up to 16 lightstyles per face and a lot more switchable light styles.

+++ quakespasm/Quake/gl_model.h
@@ -162,7 +162,7 @@ typedef struct msurface_s
 		// int is 32 bits, need an array for MAX_DLIGHTS > 32
 
 	int			lightmaptexturenum;
-	byte		styles[MAXLIGHTMAPS];
+	unsigned short		styles[MAXLIGHTMAPS];
 	int			cached_light[MAXLIGHTMAPS];	// values currently used in lightmap
 	qboolean	cached_dlight;				// true if dynamic light in cache
 	byte		*samples;		// [numstyles*surfsize]
